use doubles instead of floats, rounding issues
that I don't have a great understanding of 720 500 was coming through as 720 499.123 or whatever. moving to doubles pushes that problem back a bit. I dunno how we'll solve that.

+++ Classes/Util.cpp
@@ -77,7 +77,7 @@ std::vector<std::string> suffixes = {
     "QiD"
 };
 
-std::string _humanize_number(float value)
+std::string _humanize_number(double value)
 {
     float base = 0;
     std::string suffix = suffixes.at(0);
@@ -102,11 +102,11 @@ std::string _humanize_number(float value)
     }
 
     //for 1 567 890, we've got 1.5678 so we go to 1 568 then...
-    int embiggened = (int)std::round(value * 1000.0f);
+    int embiggened = (int)std::round(value * 1000.0);
     //... we divide by 1000 to get the nice 1.568 number and add the suffix
     // the problem is that if we've got 1.0f, it turns into 1.000 instead of 1, so I need to clear the empty 0s
     std::stringstream spss;
-    spss << std::fixed << std::setprecision(3) << embiggened / 1000.0f;
+    spss << std::fixed << std::setprecision(3) << embiggened / 1000.0;
     std::string str = spss.str();
     str.erase ( str.find_last_not_of('0') + 1, std::string::npos ); //rstrip zeroes
 
@@ -119,7 +119,7 @@ std::string _humanize_number(float value)
     return str + suffix;
 }
 
-std::string beautify_float(float value)
+std::string beautify_float(double value)
 {
     bool is_negative = value < 0;
     std::string decimal = "";
