+++ Classes/Util.h
@@ -34,7 +34,7 @@ typename mapT::mapped_type map_get(mapT& input_map, typename mapT::key_type key,
 
 std::vector<std::string>& split(const std::string &s, char delim, std::vector<std::string> &elems);
 std::vector<std::string> split(const std::string &s, char delim);
-std::string beautify_float(float value);
+std::string beautify_float(double value);
 void test_beautify_float();
 
 #endif
